Implement q_delete_dup
Ensure all duplicate nodes are removed from the list. Check if head
is valid and not empty using list_empty. Iterate through the list
with list_for_each_safe. Compare each node with the next one, and if
they are the same, delete the current node and mark it as a duplicate.
In the next iteration, remove the remaining duplicate to ensure all
identical nodes are removed.

Change-Id: If69fe7797d153b9be4b03316ce50edfa91a48dee